pyproject.toml
[project]
name = "bsgateway"
version = "0.1.0"
description = "LLM Gateway for cost-optimized routing across local, OpenAI, and Anthropic models"
requires-python = ">=3.11"
dependencies = [
    "litellm[proxy]>=1.60.0",
    "pydantic-settings>=2.0.0",
    "structlog>=23.0.0",
    "pyyaml>=6.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "cryptography>=42.0.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "bsvibe-auth[fastapi] @ git+https://github.com/blas1n/BSVibe-Auth.git@v0.1.0#subdirectory=python",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["bsgateway/tests", "tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "RUF"]
ignore = [
    "B008",   # Depends() in default args is idiomatic FastAPI
    "B017",   # pytest.raises(Exception) is fine in tests
]

[tool.coverage.run]
source = ["bsgateway"]
